bugfix in Curve initiation

+++ signal_tools.py
@@ -545,8 +545,8 @@ class Curve:
 
         elif isinstance(xy, tuple) and len(xy[0]) == len(xy[1]):
             self._check_if_sorted_and_valid(tuple(xy[0]))
-            setattr(self, "_x", np.array(xy[0]))
-            setattr(self, "_y", np.array(xy[1]))
+            setattr(self, "_x", np.array(xy[0], dtype=float))
+            setattr(self, "_y", np.array(xy[1], dtype=float))
             setattr(self, "_xy", np.vstack([self._x, self._y]))
 
         elif isinstance(xy, str):  # only works with frequencies in index (shape=(*, 2))
